# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bitunix Trader API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PositionSide = typeof PositionSide[keyof typeof PositionSide];


export const PositionSide = {
  long: 'long',
  short: 'short',
} as const;

export interface Position {
  id: string;
  symbol: string;
  side: PositionSide;
  size: number;
  entryPrice: number;
  markPrice: number;
  liquidationPrice: number;
  unrealizedPnl: number;
  unrealizedPnlPct: number;
  leverage: number;
  margin: number;
  createdAt: string;
}

export type ClosePositionRequestCloseType = typeof ClosePositionRequestCloseType[keyof typeof ClosePositionRequestCloseType];


export const ClosePositionRequestCloseType = {
  market: 'market',
  limit: 'limit',
  stop_market: 'stop_market',
  stop_limit: 'stop_limit',
  trailing_stop: 'trailing_stop',
} as const;

export interface ClosePositionRequest {
  closeType: ClosePositionRequestCloseType;
  price?: number;
  stopPrice?: number;
  trailingOffset?: number;
}

export type ReducePositionRequestCloseType = typeof ReducePositionRequestCloseType[keyof typeof ReducePositionRequestCloseType];


export const ReducePositionRequestCloseType = {
  market: 'market',
  limit: 'limit',
} as const;

export interface ReducePositionRequest {
  reducePercent: number;
  closeType: ReducePositionRequestCloseType;
  price?: number;
}

export interface OrderResult {
  orderId: string;
  status: string;
  message?: string;
}

export interface Order {
  id: string;
  symbol: string;
  side: string;
  type: string;
  price?: number;
  size: number;
  filled: number;
  status: string;
  createdAt: string;
}

export interface Trade {
  id: string;
  symbol: string;
  side: string;
  size: number;
  entryPrice: number;
  exitPrice: number;
  realizedPnl: number;
  closedAt: string;
}

export interface Account {
  totalBalance: number;
  availableBalance: number;
  usedMargin: number;
  unrealizedPnl: number;
  equity: number;
  marginRatio: number;
}

export type TriggerTriggerType = typeof TriggerTriggerType[keyof typeof TriggerTriggerType];


export const TriggerTriggerType = {
  price_target: 'price_target',
  pnl_percent: 'pnl_percent',
  trailing_stop: 'trailing_stop',
  time_based: 'time_based',
} as const;

export type TriggerCondition = typeof TriggerCondition[keyof typeof TriggerCondition];


export const TriggerCondition = {
  above: 'above',
  below: 'below',
} as const;

export type TriggerCloseType = typeof TriggerCloseType[keyof typeof TriggerCloseType];


export const TriggerCloseType = {
  market: 'market',
  limit: 'limit',
} as const;

export interface Trigger {
  id: string;
  positionId: string;
  symbol: string;
  triggerType: TriggerTriggerType;
  condition: TriggerCondition;
  value: number;
  closeType: TriggerCloseType;
  active: boolean;
  createdAt: string;
}

export type CreateTriggerRequestTriggerType = typeof CreateTriggerRequestTriggerType[keyof typeof CreateTriggerRequestTriggerType];


export const CreateTriggerRequestTriggerType = {
  price_target: 'price_target',
  pnl_percent: 'pnl_percent',
  trailing_stop: 'trailing_stop',
  time_based: 'time_based',
} as const;

export type CreateTriggerRequestCondition = typeof CreateTriggerRequestCondition[keyof typeof CreateTriggerRequestCondition];


export const CreateTriggerRequestCondition = {
  above: 'above',
  below: 'below',
} as const;

export type CreateTriggerRequestCloseType = typeof CreateTriggerRequestCloseType[keyof typeof CreateTriggerRequestCloseType];


export const CreateTriggerRequestCloseType = {
  market: 'market',
  limit: 'limit',
} as const;

export interface CreateTriggerRequest {
  positionId: string;
  symbol: string;
  triggerType: CreateTriggerRequestTriggerType;
  condition: CreateTriggerRequestCondition;
  value: number;
  closeType: CreateTriggerRequestCloseType;
}

export interface Settings {
  apiKey: string;
  apiSecret: string;
  baseUrl: string;
  coolifyEndpoint: string;
  refreshInterval: number;
  confirmCloses: boolean;
  hapticFeedback: boolean;
}

export interface SettingsUpdate {
  apiKey?: string;
  apiSecret?: string;
  baseUrl?: string;
  coolifyEndpoint?: string;
  refreshInterval?: number;
  confirmCloses?: boolean;
  hapticFeedback?: boolean;
}

export type CloseAllPositions200 = {
  closed: number;
  errors: number;
};

export type GetTradeHistoryParams = {
limit?: number;
};

export type DeleteTrigger200 = {
  success: boolean;
};