# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bitunix Trader API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: positions
    description: Position management
  - name: orders
    description: Order management
  - name: account
    description: Account info
  - name: triggers
    description: Auto-close triggers
  - name: settings
    description: App settings
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /positions:
    get:
      operationId: getPositions
      tags: [positions]
      summary: Get all open positions
      responses:
        "200":
          description: List of open positions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Position"

  /positions/{positionId}/close:
    post:
      operationId: closePosition
      tags: [positions]
      summary: Close a position
      parameters:
        - name: positionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClosePositionRequest"
      responses:
        "200":
          description: Close order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResult"

  /positions/{positionId}/reduce:
    post:
      operationId: reducePosition
      tags: [positions]
      summary: Partially reduce a position
      parameters:
        - name: positionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReducePositionRequest"
      responses:
        "200":
          description: Reduce order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResult"

  /positions/close-all:
    post:
      operationId: closeAllPositions
      tags: [positions]
      summary: Emergency close all positions at market
      responses:
        "200":
          description: All positions closed
          content:
            application/json:
              schema:
                type: object
                properties:
                  closed:
                    type: integer
                  errors:
                    type: integer
                required: [closed, errors]

  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: Get open orders
      responses:
        "200":
          description: List of open orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

  /orders/{orderId}:
    delete:
      operationId: cancelOrder
      tags: [orders]
      summary: Cancel an open order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResult"

  /history:
    get:
      operationId: getTradeHistory
      tags: [orders]
      summary: Get recent trade history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Trade history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trade"

  /account:
    get:
      operationId: getAccount
      tags: [account]
      summary: Get account balance and summary
      responses:
        "200":
          description: Account info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"

  /triggers:
    get:
      operationId: getTriggers
      tags: [triggers]
      summary: Get all auto-close triggers
      responses:
        "200":
          description: List of triggers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trigger"
    post:
      operationId: createTrigger
      tags: [triggers]
      summary: Create an auto-close trigger
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTriggerRequest"
      responses:
        "200":
          description: Trigger created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trigger"

  /triggers/{triggerId}:
    delete:
      operationId: deleteTrigger
      tags: [triggers]
      summary: Delete a trigger
      parameters:
        - name: triggerId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Trigger deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required: [success]

  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get app settings
      responses:
        "200":
          description: Current settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
    put:
      operationId: updateSettings
      tags: [settings]
      summary: Update app settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsUpdate"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    Position:
      type: object
      properties:
        id:
          type: string
        symbol:
          type: string
        side:
          type: string
          enum: [long, short]
        size:
          type: number
        entryPrice:
          type: number
        markPrice:
          type: number
        liquidationPrice:
          type: number
        unrealizedPnl:
          type: number
        unrealizedPnlPct:
          type: number
        leverage:
          type: integer
        margin:
          type: number
        createdAt:
          type: string
      required: [id, symbol, side, size, entryPrice, markPrice, liquidationPrice, unrealizedPnl, unrealizedPnlPct, leverage, margin, createdAt]

    ClosePositionRequest:
      type: object
      properties:
        closeType:
          type: string
          enum: [market, limit, stop_market, stop_limit, trailing_stop]
        price:
          type: number
        stopPrice:
          type: number
        trailingOffset:
          type: number
      required: [closeType]

    ReducePositionRequest:
      type: object
      properties:
        reducePercent:
          type: number
        closeType:
          type: string
          enum: [market, limit]
        price:
          type: number
      required: [reducePercent, closeType]

    OrderResult:
      type: object
      properties:
        orderId:
          type: string
        status:
          type: string
        message:
          type: string
      required: [orderId, status]

    Order:
      type: object
      properties:
        id:
          type: string
        symbol:
          type: string
        side:
          type: string
        type:
          type: string
        price:
          type: number
        size:
          type: number
        filled:
          type: number
        status:
          type: string
        createdAt:
          type: string
      required: [id, symbol, side, type, size, filled, status, createdAt]

    Trade:
      type: object
      properties:
        id:
          type: string
        symbol:
          type: string
        side:
          type: string
        size:
          type: number
        entryPrice:
          type: number
        exitPrice:
          type: number
        realizedPnl:
          type: number
        closedAt:
          type: string
      required: [id, symbol, side, size, entryPrice, exitPrice, realizedPnl, closedAt]

    Account:
      type: object
      properties:
        totalBalance:
          type: number
        availableBalance:
          type: number
        usedMargin:
          type: number
        unrealizedPnl:
          type: number
        equity:
          type: number
        marginRatio:
          type: number
      required: [totalBalance, availableBalance, usedMargin, unrealizedPnl, equity, marginRatio]

    Trigger:
      type: object
      properties:
        id:
          type: string
        positionId:
          type: string
        symbol:
          type: string
        triggerType:
          type: string
          enum: [price_target, pnl_percent, trailing_stop, time_based]
        condition:
          type: string
          enum: [above, below]
        value:
          type: number
        closeType:
          type: string
          enum: [market, limit]
        active:
          type: boolean
        createdAt:
          type: string
      required: [id, positionId, symbol, triggerType, condition, value, closeType, active, createdAt]

    CreateTriggerRequest:
      type: object
      properties:
        positionId:
          type: string
        symbol:
          type: string
        triggerType:
          type: string
          enum: [price_target, pnl_percent, trailing_stop, time_based]
        condition:
          type: string
          enum: [above, below]
        value:
          type: number
        closeType:
          type: string
          enum: [market, limit]
      required: [positionId, symbol, triggerType, condition, value, closeType]

    Settings:
      type: object
      properties:
        apiKey:
          type: string
        apiSecret:
          type: string
        baseUrl:
          type: string
        coolifyEndpoint:
          type: string
        refreshInterval:
          type: integer
        confirmCloses:
          type: boolean
        hapticFeedback:
          type: boolean
      required: [apiKey, apiSecret, baseUrl, coolifyEndpoint, refreshInterval, confirmCloses, hapticFeedback]

    SettingsUpdate:
      type: object
      properties:
        apiKey:
          type: string
        apiSecret:
          type: string
        baseUrl:
          type: string
        coolifyEndpoint:
          type: string
        refreshInterval:
          type: integer
        confirmCloses:
          type: boolean
        hapticFeedback:
          type: boolean