# lib/api-zod/src/generated/types/account.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bitunix Trader API
 * OpenAPI spec version: 0.1.0
 */

export interface Account {
  totalBalance: number;
  availableBalance: number;
  usedMargin: number;
  unrealizedPnl: number;
  equity: number;
  marginRatio: number;
}