# lib/api-zod/src/generated/types/closePositionRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bitunix Trader API
 * OpenAPI spec version: 0.1.0
 */
import type { ClosePositionRequestCloseType } from './closePositionRequestCloseType';

export interface ClosePositionRequest {
  closeType: ClosePositionRequestCloseType;
  price?: number;
  stopPrice?: number;
  trailingOffset?: number;
}