# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bitunix Trader API
 * OpenAPI spec version: 0.1.0
 */

export interface Order {
  id: string;
  symbol: string;
  side: string;
  type: string;
  price?: number;
  size: number;
  filled: number;
  status: string;
  createdAt: string;
}