# lib/api-zod/src/generated/types/settings.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bitunix Trader API
 * OpenAPI spec version: 0.1.0
 */

export interface Settings {
  apiKey: string;
  apiSecret: string;
  baseUrl: string;
  coolifyEndpoint: string;
  refreshInterval: number;
  confirmCloses: boolean;
  hapticFeedback: boolean;
}