# lib/api-zod/src/generated/types/trigger.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bitunix Trader API
 * OpenAPI spec version: 0.1.0
 */
import type { TriggerCloseType } from './triggerCloseType';
import type { TriggerCondition } from './triggerCondition';
import type { TriggerTriggerType } from './triggerTriggerType';

export interface Trigger {
  id: string;
  positionId: string;
  symbol: string;
  triggerType: TriggerTriggerType;
  condition: TriggerCondition;
  value: number;
  closeType: TriggerCloseType;
  active: boolean;
  createdAt: string;
}